+++ common/config/babel/.babelrc.js
@@ -36,7 +36,7 @@ process.env['COMMUNICATION_REACT_FLAVOR'] === 'stable' &&
         // rooms
         'rooms',
         // Add people dropdown and dialpad modal in people pane
-        'PeoplePaneDropdown',
+        'people-pane-dropdown',
       ],
       // A list of stabilized features.
       // These features can be listed in the conditional compilation directives without
