+++ packages/react-composites/src/composites/CallWithChatComposite/components/PreparedDialpad.styles.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import { IButtonStyles, IModalStyles, Theme } from '@fluentui/react';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
+/* @conditional-compile-remove(people-pane-dropdown) */
 import { DialpadStyles } from '@internal/react-components';
 
 /**
@@ -15,7 +15,7 @@ export const themedDialpadModelStyle = (theme: Theme): Partial<IModalStyles> => 
   }
 });
 
-/* @conditional-compile-remove(PeoplePaneDropdown) */
+/* @conditional-compile-remove(people-pane-dropdown) */
 /**
  * @private
  */
