+++ packages/react-composites/src/composites/CallWithChatComposite/components/PreparedDialpad.tsx
@@ -1,20 +1,20 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 import React from 'react';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
+/* @conditional-compile-remove(people-pane-dropdown) */
 import { useState } from 'react';
 import { useMemo } from 'react';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
+/* @conditional-compile-remove(people-pane-dropdown) */
 import { Dialpad, DialpadStyles } from '@internal/react-components';
 import { _DrawerMenu, _DrawerMenuItemProps, _DrawerSurface } from '@internal/react-components';
 import { IModalStyles, Modal, Stack, useTheme, Text, IconButton } from '@fluentui/react';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
+/* @conditional-compile-remove(people-pane-dropdown) */
 import { IButtonStyles, PrimaryButton } from '@fluentui/react';
 import { drawerContainerStyles } from '../styles/CallWithChatCompositeStyles';
 import { themedDialpadModelStyle } from './PreparedDialpad.styles';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
+/* @conditional-compile-remove(people-pane-dropdown) */
 import { themedCallButtonStyle, themedDialpadStyle } from './PreparedDialpad.styles';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
+/* @conditional-compile-remove(people-pane-dropdown) */
 import { CallWithChatCompositeIcon } from '../../common/icons';
 
 /** @private */
@@ -36,18 +36,18 @@ export interface PreparedDialpadProps {
 /** @private */
 export const PreparedDialpad = (props: PreparedDialpadProps): JSX.Element => {
   const { strings, isMobile, showDialpad, onDismissDialpad } = props;
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
+  /* @conditional-compile-remove(people-pane-dropdown) */
   const [textFieldInput, setTextFieldInput] = useState('');
 
   const theme = useTheme();
 
   const onDismissTriggered = (): void => {
-    /* @conditional-compile-remove(PeoplePaneDropdown) */
+    /* @conditional-compile-remove(people-pane-dropdown) */
     setTextFieldInput('');
     onDismissDialpad();
   };
 
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
+  /* @conditional-compile-remove(people-pane-dropdown) */
   const onClickCall = (): void => {
     //place holder for adding calling functionality
     console.log(textFieldInput);
@@ -55,20 +55,20 @@ export const PreparedDialpad = (props: PreparedDialpadProps): JSX.Element => {
 
   const dialpadModelStyle: Partial<IModalStyles> = useMemo(() => themedDialpadModelStyle(theme), [theme]);
 
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
+  /* @conditional-compile-remove(people-pane-dropdown) */
   const dialpadStyle: Partial<DialpadStyles> = useMemo(() => themedDialpadStyle(isMobile, theme), [theme, isMobile]);
 
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
+  /* @conditional-compile-remove(people-pane-dropdown) */
   const callButtonStyle: Partial<IButtonStyles> = useMemo(() => themedCallButtonStyle(theme), [theme]);
 
   const dialpadComponent = (): JSX.Element => {
-    /* @conditional-compile-remove(PeoplePaneDropdown) */
+    /* @conditional-compile-remove(people-pane-dropdown) */
     return (
       <>
         <Dialpad onChange={setTextFieldInput} styles={dialpadStyle} />
         <PrimaryButton
           text={strings.dialpadStartCallButtonLabel}
-          onRenderIcon={() => <CallWithChatCompositeIcon iconName="Call" />}
+          onRenderIcon={() => <CallWithChatCompositeIcon iconName="DialpadStartCall" />}
           onClick={onClickCall}
           styles={callButtonStyle}
           disabled={textFieldInput === ''}
@@ -81,7 +81,7 @@ export const PreparedDialpad = (props: PreparedDialpadProps): JSX.Element => {
 
   if (isMobile) {
     return (
-      <Stack>
+      <Stack data-ui-id="call-with-chat-composite-dialpad">
         {showDialpad && (
           <Stack styles={drawerContainerStyles}>
             <_DrawerSurface onLightDismiss={onDismissTriggered}>
@@ -102,6 +102,7 @@ export const PreparedDialpad = (props: PreparedDialpadProps): JSX.Element => {
           onDismiss={onDismissTriggered}
           isBlocking={true}
           styles={dialpadModelStyle}
+          data-ui-id="call-with-chat-composite-dialpad"
         >
           <Stack horizontal horizontalAlign="space-between" verticalAlign="center">
             <Text style={{ fontWeight: 600 }}>{strings.dialpadModalTitle}</Text>
