+++ packages/react-composites/src/composites/common/PeoplePaneContent.tsx
@@ -10,8 +10,6 @@ import {
   Stack,
   useTheme
 } from '@fluentui/react';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
-import { IContextualMenuProps, IContextualMenuStyles } from '@fluentui/react';
 import {
   ParticipantList,
   ParticipantListParticipant,
@@ -19,13 +17,8 @@ import {
   ParticipantMenuItemsCallback,
   _DrawerMenuItemProps
 } from '@internal/react-components';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
-import { _DrawerMenu, _DrawerSurface } from '@internal/react-components';
 import copy from 'copy-to-clipboard';
 import React, { useMemo } from 'react';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
-import { useState } from 'react';
-import { CallAdapter } from '../CallComposite';
 import { CallWithChatCompositeStrings } from '../CallWithChatComposite';
 import { usePropsFor } from '../CallComposite/hooks/usePropsFor';
 import { AvatarPersonaDataCallback } from '../common/AvatarPersona';
@@ -40,15 +33,9 @@ import {
   participantListContainerStyles,
   peoplePaneContainerStyle
 } from './styles/PeoplePaneContent.styles';
-/* @conditional-compile-remove(one-to-n-calling) */
-import { CallCompositeStrings } from '../CallComposite';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
-import { themedMenuStyle } from './styles/PeoplePaneContent.styles';
 import { themedCopyLinkButtonStyles } from './styles/PeoplePaneContent.styles';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
-import { drawerContainerStyles } from '../CallWithChatComposite/styles/CallWithChatCompositeStyles';
-/* @conditional-compile-remove(PeoplePaneDropdown) */
-import { PreparedDialpad } from '../CallWithChatComposite/components/PreparedDialpad';
+import { AddPeopleDropdown } from '../CallWithChatComposite/components/AddPeopleDropdown';
+import { convertContextualMenuItemToDrawerMenuItem } from '../CallWithChatComposite/ConvertContextualMenuItemToDrawerMenuItem';
 
 /**
  * @private
@@ -124,101 +111,10 @@ export const PeoplePaneContent = (props: {
     [props.mobileView, theme]
   );
 
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
-  // desktop add people button with context menu
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
-  const [showDialpad, setShowDialpad] = useState(false);
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
-  const menuStyleThemed: Partial<IContextualMenuStyles> = useMemo(() => themedMenuStyle(theme), [theme]);
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
-  const defaultMenuProps = useMemo((): IContextualMenuProps => {
-    const menuProps: IContextualMenuProps = {
-      styles: menuStyleThemed,
-      items: [],
-      useTargetWidth: true,
-      calloutProps: {
-        // Disable dismiss on resize to work around a couple Fluent UI bugs
-        // - The Callout is dismissed whenever *any child of window (inclusive)* is resized. In practice, this
-        //   happens when we change the VideoGallery layout, or even when the video stream element is internally resized
-        //   by the headless SDK.
-        // - There is a `preventDismissOnEvent` prop that we could theoretically use to only dismiss when the target of
-        //   of the 'resize' event is the window itself. But experimentation shows that setting that prop doesn't
-        //   deterministically avoid dismissal.
-        //
-        // A side effect of this workaround is that the context menu stays open when window is resized, and may
-        // get detached from original target visually. That bug is preferable to the bug when this value is not set -
-        // The Callout (frequently) gets dismissed automatically.
-        preventDismissOnResize: true
-      }
-    };
-
-    if (inviteLink) {
-      menuProps.items.push({
-        key: 'InviteLinkKey',
-        text: strings.copyInviteLinkButtonLabel,
-        itemProps: { styles: copyLinkButtonStylesThemed },
-        iconProps: { iconName: 'Link', style: linkIconStyles },
-        onClick: () => copy(inviteLink)
-      });
-    }
-
-    menuProps.items.push({
-      key: 'DialpadKey',
-      text: strings.openDialpadButtonLabel,
-      itemProps: { styles: copyLinkButtonStylesThemed },
-      iconProps: { iconName: 'Dialpad', style: linkIconStyles },
-      onClick: () => setShowDialpad(true)
-    });
-
-    return menuProps;
-  }, [
-    strings.copyInviteLinkButtonLabel,
-    strings.openDialpadButtonLabel,
-    copyLinkButtonStylesThemed,
-    inviteLink,
-    menuStyleThemed
-  ]);
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
-  const onDismissDialpad = (): void => {
-    setShowDialpad(false);
-  };
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
-  //mobile add people button with botten sheet drawers
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
-  const [addPeopleDrawerMenuItems, setAddPeopleDrawerMenuItems] = useState<_DrawerMenuItemProps[]>([]);
-  /* @conditional-compile-remove(PeoplePaneDropdown) */
-  const setDrawerMenuItemsForAddPeople: () => void = useMemo(() => {
-    return () => {
-      const drawerMenuItems = defaultMenuProps.items.map((contextualMenu: IContextualMenuItem) =>
-        convertContextualMenuItemToDrawerMenuItem(contextualMenu, () => setAddPeopleDrawerMenuItems([]))
-      );
-      setAddPeopleDrawerMenuItems(drawerMenuItems);
-    };
-  }, [defaultMenuProps, setAddPeopleDrawerMenuItems]);
-
   // for conditional compile
   const mobileViewAddPeopleButton = (inviteLink?: string): JSX.Element => {
-    /* @conditional-compile-remove(PeoplePaneDropdown) */
-    return (
-      <Stack>
-        <Stack.Item styles={copyLinkButtonContainerStyles}>
-          <PrimaryButton
-            onClick={setDrawerMenuItemsForAddPeople}
-            styles={copyLinkButtonStylesThemed}
-            onRenderIcon={() => <CallWithChatCompositeIcon iconName="PeoplePaneAddPerson" />}
-            text={strings.peoplePaneAddPeopleButtonLabel}
-          />
-        </Stack.Item>
-
-        {addPeopleDrawerMenuItems.length > 0 && (
-          <Stack styles={drawerContainerStyles}>
-            <_DrawerMenu onLightDismiss={() => setAddPeopleDrawerMenuItems([])} items={addPeopleDrawerMenuItems} />
-          </Stack>
-        )}
-
-        <PreparedDialpad isMobile strings={strings} showDialpad={showDialpad} onDismissDialpad={onDismissDialpad} />
-      </Stack>
-    );
+    /* @conditional-compile-remove(people-pane-dropdown) */
+    return <AddPeopleDropdown strings={strings} mobileView={props.mobileView} inviteLink={inviteLink} />;
 
     return (
       <Stack>
@@ -237,27 +133,11 @@ export const PeoplePaneContent = (props: {
   };
 
   const desktopViewAddPeopleButton = (inviteLink?: string): JSX.Element => {
-    /* @conditional-compile-remove(PeoplePaneDropdown) */
+    /* @conditional-compile-remove(people-pane-dropdown) */
     return (
-      <Stack>
-        <PreparedDialpad
-          isMobile={false}
-          strings={strings}
-          showDialpad={showDialpad}
-          onDismissDialpad={onDismissDialpad}
-        />
-        <Stack tokens={peoplePaneContainerTokens}>
-          <Stack styles={copyLinkButtonStackStyles}>
-            <DefaultButton
-              onRenderIcon={() => <CallWithChatCompositeIcon iconName="PeoplePaneAddPerson" />}
-              text={strings.peoplePaneAddPeopleButtonLabel}
-              menuProps={defaultMenuProps}
-              styles={copyLinkButtonStylesThemed}
-            />
-          </Stack>
-
-          {participantList}
-        </Stack>
+      <Stack tokens={peoplePaneContainerTokens}>
+        <AddPeopleDropdown strings={strings} mobileView={props.mobileView} inviteLink={inviteLink} />
+        {participantList}
       </Stack>
     );
 
@@ -323,26 +203,4 @@ const createDefaultContextualMenuItems = (
     });
   }
   return menuItems;
-};
-
-/**
- * Convert IContextualMenuItem to _DrawerMenuItemProps
- * @param contextualMenu - IContextualMenuItem to convert
- * @param onDrawerMenuItemClick - callback to call when converted DrawerMenuItem is clicked
- * @returns DrawerMenuItem
- */
-const convertContextualMenuItemToDrawerMenuItem = (
-  contextualMenu: IContextualMenuItem,
-  onDrawerMenuItemClick: () => void
-): _DrawerMenuItemProps => {
-  return {
-    itemKey: contextualMenu.key,
-    onItemClick: () => {
-      contextualMenu.onClick?.();
-      onDrawerMenuItemClick();
-    },
-    iconProps: contextualMenu.iconProps,
-    text: contextualMenu.text,
-    disabled: contextualMenu.disabled
-  };
 };
